refactor: remove compositor_key_repeat feature flag, use Option-based loop_handle
Replace the feature flag with an opt-in trait method that returns Option.
Compositors enable key repeat by overriding SeatHandler::loop_handle()
to return Some(handle). Move manage_key_repeat into input/keyboard/mod.rs
since it no longer depends on any wayland-specific types.
