+++ anvil/Cargo.toml
@@ -23,7 +23,7 @@ profiling = { version = "1.0" }
 
 [dependencies.smithay]
 default-features = false
-features = ["desktop", "wayland_frontend", "compositor_key_repeat"]
+features = ["desktop", "wayland_frontend"]
 path = ".."
 
 [dependencies.x11rb]
